days/069-072-django-rest/demo/requirements.txt
django
django-registration
djangorestframework
requests
django-rest-swagger
